add failing test - when cursor changes to non-cursor
Add a failing test for when a cursor changes to a non-cursor, shouldComponentUpdate throws an error

    TypeError: Cannot read property 'deref' of undefined

For the test I used a named property other than cursor since the test has some filtering for the special options.cursor in which is assigns the property cursor or not and hides the issue.

shouldComponentUpdate should work properly regardless of the property name, so if a property is a cursor and then by the same name changes to non-cursor we should not throw a type error.