hasChangedCursor handle cursor to non-cursor change
shouldComponentUpdate's hasChangedCursor should handle the case where a property once was a cursor but now has changed to a non-cursor.

Previously the isEqualCursor method would throw a TypeError when unCursor tries to access property deref of an undefined.

Since hasChangedCursor filters the properties to check for cursors, we have to handle the case where after the filtering, a property is undefined and we can short circuit the shouldComponentUpdate with true.